requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asn1crypto==1.5.1
async-timeout==4.0.3
cbor2==5.6.4
certifi==2024.8.30
certvalidator==0.11.1
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
exceptiongroup==1.2.2
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
oscrypto==1.3.0
py-webauthn==0.0.4
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
